docker error fix
